added a symbolic module containing memory & model
