+++ src/see/mod.rs
@@ -16,8 +16,8 @@ use crate::cfg::{generate_cfg, generate_dot_cfg};
 use crate::shared::Config;
 use crate::shared::ExitCode;
 use crate::shared::{panic_with_diagnostics, Diagnostics, Error};
-use crate::sym_model::PathConstraints;
-use crate::sym_model::{ReferenceValue, SymExpression, SymMemory, SymValue};
+use crate::symbolic::model::{PathConstraints, ReferenceValue, SymExpression, SymValue};
+use crate::symbolic::memory::{SymMemory};
 use crate::z3::build_ctx;
 use petgraph::graph::NodeIndex;
 use petgraph::visit::EdgeRef;
