+++ src/symbolic/memory.rs
@@ -1,4 +1,4 @@
-//! Intermediate types & utilities to encode & manipulate ast before we call z3
+//! Symbolic model representing the stack and heap while symbolically executing a program
 //!
 use rustc_hash::FxHashMap;
 use ::z3::Context;
@@ -9,95 +9,8 @@ use crate::ast::*;
 use crate::shared::{panic_with_diagnostics, Error, Scope};
 use crate::z3;
 
-//----------------------//
-// Symbolic expressions //
-//----------------------//
+use super::model::{SymExpression, Reference, ReferenceValue, Substituted, SymValue};
 
-#[derive(Clone)]
-pub struct PathConstraints {
-    constraints: Vec<PathConstraint>,
-}
-
-#[derive(Clone)]
-pub enum PathConstraint {
-    Assert(Substituted),
-    Assume(Substituted),
-}
-
-impl Default for PathConstraints {
-    fn default() -> Self {
-        PathConstraints {
-            constraints: vec![],
-        }
-    }
-}
-
-impl PathConstraints {
-    pub fn combine<'a>(&'a self) -> Substituted {
-        let mut constraints = Expression::Literal(Literal::Boolean(true));
-
-        for constraint in self.constraints.iter().rev() {
-            match constraint {
-                PathConstraint::Assert(Substituted(assertion)) => {
-                    constraints =
-                        Expression::And(Box::new(assertion.clone()), Box::new(constraints));
-                }
-                PathConstraint::Assume(Substituted(assumption)) => {
-                    constraints =
-                        Expression::Implies(Box::new(assumption.clone()), Box::new(constraints));
-                }
-            }
-        }
-        return Substituted(constraints);
-    }
-
-    /// adds a new constraint
-    pub fn push_assertion(&mut self, s: Substituted) {
-        self.constraints.push(PathConstraint::Assert(s));
-    }
-    /// adds a new constraint
-    pub fn push_assumption(&mut self, s: Substituted) {
-        self.constraints.push(PathConstraint::Assume(s));
-    }
-}
-
-/// containertype to indicate substitution of variables has already occured on expression
-#[derive(Debug, Clone)]
-pub struct Substituted(pub Expression);
-
-/// Containertype containing normalized expression. 
-/// Given 3 expressions `e1 = 1+2+3`, `e2 = 3+2+1` and `e1 = 2+3+1` 
-/// then `Normalized(e1) == Normalized(e2)`, `Normalized(e2) == Normalized(e3)` and so on
-pub struct Normalized(pub Substituted);
-
-pub type Reference = Uuid;
-
-#[derive(Debug, Clone)]
-pub enum SymValue {
-    Uninitialized,
-    Expr(Substituted),
-}
-
-#[derive(Debug, Clone)]
-pub enum SymExpression {
-    Int(SymValue),
-    Bool(SymValue),
-    Ref((Type, Reference)),
-}
-
-/// Consists of `identifier` (= classname) and a hashmap describing it's fields
-pub type Object = (Identifier, FxHashMap<Identifier, (Type, SymExpression)>);
-
-/// Consists of type, concrete array and Substituted expression representing length
-pub type Array = (Type, Vec<SymExpression>, Substituted);
-
-#[derive(Debug, Clone)]
-pub enum ReferenceValue {
-    Object(Object),
-    Array(Array),
-    /// Takes classname as input
-    UninitializedObj(Class),
-}
 
 //-----------------//
 // Symbolic memory //
@@ -113,7 +26,6 @@ type SymStack<'a> = Vec<Frame<'a>>;
 
 type SymHeap = FxHashMap<Reference, ReferenceValue>;
 
-type Test = FxHashMap<Substituted, Substituted>;
 
 #[derive(Clone)]
 pub struct SymMemory<'ctx> {
@@ -277,7 +189,7 @@ impl<'a> SymMemory<'a> {
 
 
         //get array type, array and length
-        let (l_expr, r_expr, length)= match self.stack_get(&arr_name){
+        let (_, _, length)= match self.stack_get(&arr_name){
             Some(SymExpression::Ref((_, r))) => match self.heap.get(&r){
                 Some(ReferenceValue::Array((ty, arr, length))) => (ty, arr, length),
                 otherwise => panic_with_diagnostics(&format!("{:?} is not an array and can't be assigned to in assignment '{}[{:?}] := {:?}'", otherwise, arr_name, subt_index, var), &self),
@@ -307,7 +219,7 @@ impl<'a> SymMemory<'a> {
     pub fn heap_get_length(&self, arr_name: &Identifier) -> SymExpression {
         match self.stack_get(&arr_name){
         Some(SymExpression::Ref((_, r))) => match self.heap.get(&r){
-            Some(ReferenceValue::Array((l_expr, r_expr, length))) => SymExpression::Int(SymValue::Expr(length.clone())),
+            Some(ReferenceValue::Array((_, _, length))) => SymExpression::Int(SymValue::Expr(length.clone())),
             otherwise => panic_with_diagnostics(&format!("Can't return length of {} since the value it references to ({:?}) is not an array", arr_name, otherwise), &self),
         },
         _ => panic_with_diagnostics(&format!("{} is not a reference", arr_name), &self),
